fix(test): move SIGHUP tests to separate file for Windows compatibility
Use build tag instead of runtime skip to exclude Unix-specific
SIGHUP signal tests from Windows builds.
